--- frontend/src/components/templates/TargetApi.js
@@ -1,100 +1,0 @@
-import React from 'react';
-import Form from '../../components/templates/Form';
-import OrganisationSection from '../../components/organisms/form-sections/OrganisationSection';
-import DemarcheSection from '../../components/organisms/form-sections/DemarcheSection';
-import DescriptionSection from '../../components/organisms/form-sections/DescriptionSection';
-import DonneesSection from '../../components/organisms/form-sections/DonneesSection';
-import CadreJuridiqueSection from '../../components/organisms/form-sections/CadreJuridiqueSection';
-import CguSection from '../../components/organisms/form-sections/CguSection';
-import ÉquipeSection from '../../components/organisms/form-sections/ÉquipeSection';
-import { getDefaultDocumentationUrl } from '../organisms/Nav';
-import Link from '../../components/atoms/hyperTexts/Link';
-import { useRemark } from 'react-remark';
-
-const DonneesDescription = () => (
-  <>
-    <p>
-      La loi informatique et libertés définit les principes à respecter lors de
-      la collecte, du traitement et de la conservation de données personnelles.
-    </p>
-    <p>L’article 6 précise :</p>
-    <ul>
-      <li>
-        3° [les données] sont adéquates, pertinentes et non excessives au regard
-        des finalités pour lesquelles elles sont collectées et de leurs
-        traitements ultérieurs ;
-      </li>
-      <li>
-        4° Elles sont exactes, complètes et, si nécessaire, mises à jour ; les
-        mesures appropriées doivent être prises pour que les données inexactes
-        ou incomplètes au regard des finalités pour lesquelles elles sont
-        collectées ou traitées soient effacées ou rectifiées ;
-      </li>
-    </ul>
-    <p>
-      Nous vous remercions de sélectionner uniquement les données strictement
-      nécessaires à votre téléservice. Le non-respect du principe de
-      proportionnalité vous expose vis-à-vis de la CNIL.
-    </p>
-  </>
-);
-
-const CadreJuridiqueDescription = () => (
-  <>
-    <p>
-      Pour pouvoir bénéficier du raccordement à l’API Particulier, le cadre
-      légal et réglementaire des fournisseurs de service doit permettre à la
-      DINUM de transmettre des données personnelles à votre entité
-      administrative.
-    </p>
-    <p>
-      Dans le cas où vous représentez une collectivité, veuillez joindre la
-      délibération du conseil municipal explicitant l’usage des données
-      demandées.
-    </p>
-    <p>
-      Important : à priori, vous n’avez pas de nouvelle délibération à réaliser.
-      Il vous suffit de déposer la dernière délibération tarifaire qui détaille
-      les barèmes de facturation des services, ou le règlement qui décrit les
-      données nécessaires à une instruction.
-    </p>
-    <p>
-      Pour en savoir plus, consultez{' '}
-      <Link inline href="https://api.gouv.fr/guides/deliberation-api-part">
-        notre guide sur ce qu’est une bonne délibération
-      </Link>
-      .
-    </p>
-  </>
-);
-
-const TargetApi = ({
-  target_api,
-  email,
-  scopes,
-  editorList,
-  demarches,
-  cguLink,
-}) => (
-  <Form
-    target_api={target_api}
-    demarches={demarches}
-    contactEmail={email}
-    documentationUrl={getDefaultDocumentationUrl(target_api)}
-  >
-    <OrganisationSection editorList={editorList} />
-    <DemarcheSection availableScopes={scopes} />
-    <DescriptionSection />
-    <DonneesSection
-      availableScopes={scopes}
-      DonneesDescription={DonneesDescription}
-    />
-    <CadreJuridiqueSection
-      CadreJuridiqueDescription={CadreJuridiqueDescription}
-    />
-    <ÉquipeSection />
-    <CguSection cguLink={cguLink} />
-  </Form>
-);
-
-export default TargetApi;
